update to nextstepjs v2
