+++ src/app/docs/(v2)/nextjs/dark-mode/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import DarkMode from '@/components/docs/DarkMode';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'Next.js Dark Mode Support - NextStepjs',
