+++ src/app/docs/(v2)/nextjs/routing/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import Routing from '@/components/docs/Routing';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'Next.js Routing During Tours - NextStepjs',
