+++ src/app/docs/(v2)/react-router/localization/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import Localization from '@/components/docs/Localization';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'React Router Localization - NextStepjs',
