+++ src/app/docs/(v2)/react/viewport/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import Viewport from '@/components/docs/Viewport';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'React NextStepViewport - NextStepjs',
