+++ src/app/docs/(v2)/remix/api/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import Api from '@/components/docs/Api';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'Remix API Reference - NextStepjs',
