+++ src/app/docs/(v2)/remix/basic-setup/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import BasicSetup from '@/components/docs/BasicSetup';
-import { metaKeywords } from '@/app/layout';
+import { metaKeywords } from '@/app/metaKeywords';
 
 export const metadata: Metadata = {
   title: 'Remix Basic Setup - NextStepjs',
